+++ src/features/index.ts
@@ -5,6 +5,7 @@ import nameChecker from './nameChecker'
 import ping from './ping'
 import preventEmojiSpam from './preventEmojiSpam'
 import stickyMessage from './stickyMessage'
+import threadPruner from './threadPruner'
 
 export default [
   memberUpdateLogger,
@@ -12,4 +13,5 @@ export default [
   ping,
   preventEmojiSpam,
   stickyMessage,
+  threadPruner,
 ] as Plugin[]
